Fixes for buliding for 32bit targets:   * u32/u64/uword mismatches   * pointer-to-int fixes   * printf formatting issues   * issues with incorrect "ULL" and related suffixes   * structure alignment and padding issues

+++ src/svm/ssvm.c
@@ -177,7 +177,7 @@ map_it:
   return SSVM_API_ERROR_SLAVE_TIMEOUT;
 
 re_map_it:
-  ssvm->requested_va = (u64) sh->ssvm_va;
+  ssvm->requested_va = sh->ssvm_va;
   ssvm->ssvm_size = sh->ssvm_size;
   munmap (sh, MMAP_PAGESIZE);
